Strip down everything except the assembler, rename MojoShaderAssembler.
The original MojoShader project removed the compiler, this just does the inverse with the one feature that was fully working.

This is far from clean: internal.h and common.c need to be stripped down, and CMake needs to include MojoShader properly for the include as well as the shader parser for finderrors. But it's a start!